requirements.txt
rich>=13.0.0
huggingface-hub>=0.20.0
requests>=2.31.0
psutil>=5.9.0
Pillow>=10.0.0
mutagen>=1.47.0
